Reorganized Files & Rename
Created a pages folder that contains the main features of the app. Previously all of this was stored in components. Also renamed the "Diet" page to "Nutrition" to keep the language consistent with the project plan.
