+++ client/src/App.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import { BrowserRouter, Routes, Route, Navigate} from 'react-router-dom';
 import { useAuthContext } from './hooks/useAuthContext'
-import { Login } from './components/Login/index_login';
-import { Register } from './components/Login/index_register';
-import Dashboard from './components/Dashboard';
-import Exercise from './components/Exercise';
-import Diet from './components/Diet';
-import Sleep from './components/Sleep';
+import { Login } from './pages/Login/index_login';
+import { Register } from './pages/Login/index_register';
+import Dashboard from './pages/Dashboard';
+import Exercise from './pages/Exercise';
+import Nutrition from './pages/Nutrition';
+import Sleep from './pages/Sleep';
 import './index.css';
 
 function App() {
@@ -21,7 +21,7 @@ function App() {
           <Route path="/register" element={!user ? <Register /> : <Navigate to="/dashboard" />} />
           <Route path="/dashboard" element={user ? <Dashboard /> : <Navigate to="/login" />} />
           <Route path="/exercise" element={user ? <Exercise /> : <Navigate to="/login" />} />
-          <Route path="/diet" element={user ? <Diet /> : <Navigate to="/login" />} />
+          <Route path="/nutrition" element={user ? <Nutrition /> : <Navigate to="/login" />} />
           <Route path="/sleep" element={user ? <Sleep /> : <Navigate to="/login" />} />
         </Routes>
       </BrowserRouter>
